% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_set.R
\name{addHyperplane<-}
\alias{addHyperplane<-}
\title{Setter Function}
\usage{
addHyperplane(object) <- value
}
\arguments{
\item{object}{[POLYTOPE] object of class 'POLYTOPE'.}
}
\value{
Returns the modified input object.
}
\description{
Adding one or several inequality equations.
}
