% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{backTransform}
\alias{backTransform}
\alias{backTransform,RP-method}
\title{backTransform}
\usage{
backTransform(object)

\S4method{backTransform}{RP}(object)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{RP}}
}
\value{
Returns an object of the same class as the imput 
        but transformed to (d-1)-dimensional space.
}
\description{
Perrforms a (isometric, i.e. distance preserving) transformation
             of body from d-demensional space to (d-1)-demensional space.
             Note that the body needs to be a simplex or a the intersection of a 
             body with a simplex (i.e. a constrained simplex).
}
