% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{dimension}
\alias{dimension}
\alias{dimension,POLYTOPE-method}
\alias{dimension,}
\alias{POLYTOPE-method}
\alias{dimension,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{dimension,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{dimension}
\usage{
dimension(object)

\S4method{dimension}{POLYTOPE}(object)

\S4method{dimension}{ELLIPSOID}(object)

\S4method{dimension}{CONVEXBODY}(object)
}
\arguments{
\item{object}{[POLYTOPE] Any class that inherits from POLYTOPE.}
}
\value{
Returns a matrix.
}
\description{
Returns the dimension of an object
}
