% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{dists2Body}
\alias{dists2Body}
\alias{dists2Body,POLYTOPE-method}
\alias{dist2Body,}
\alias{POLYTOPE-method}
\alias{dists2Body,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{dists2Body,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{dists2Body}
\usage{
dists2Body(object, x, u = NULL)

\S4method{dists2Body}{POLYTOPE}(object, x, u = NULL)

\S4method{dists2Body}{ELLIPSOID}(object, x, u = NULL)

\S4method{dists2Body}{CONVEXBODY}(object, x, u = NULL)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{CONVEXBODY}, 
\code{POLYTOPE} or \code{ELLIPSOID}.}

\item{x}{[vector] Point within the body.}

\item{u}{[vector] Optional direction vector.}
}
\value{
Returns a numeric vector of length two.
}
\description{
Computes the distances from a point x to the boundary of a body in 
             in directions u and -u.
}
