% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{load a CSV file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{path to the csv file}
}
\value{
the function returns a data.frame based on the CSV file.
}
\description{
The function loads a CSV file defined by \code{filename} argument and returns
a tibble
}
\examples{
\dontrun{
accident_2014<-fars_read("./data/accident_2014.csv.bz2")
}

}
