% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getAmat}
\alias{getAmat}
\alias{getAmat,POLYTOPE-method}
\alias{getAmat,}
\alias{POLYTOPE-method}
\alias{getAmat,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{getAmat}
\usage{
getAmat(object)

\S4method{getAmat}{POLYTOPE}(object)

\S4method{getAmat}{CONVEXBODY}(object)
}
\arguments{
\item{object}{[POLYTOPE] Any class that inherits from POLYTOPE.}
}
\value{
Returns a matrix.
}
\description{
Get the A-matrix from a \code{POLYTOPE} object.
}
