% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getBoundaryPoint}
\alias{getBoundaryPoint}
\alias{getBoundaryPoint,POLYTOPE-method}
\alias{getBoundaryPoint,}
\alias{POLYTOPE-method}
\alias{getBoundaryPoint,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{getBoundaryPoint,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{getBoundaryPoint}
\usage{
getBoundaryPoint(object)

\S4method{getBoundaryPoint}{POLYTOPE}(object)

\S4method{getBoundaryPoint}{ELLIPSOID}(object)

\S4method{getBoundaryPoint}{CONVEXBODY}(object)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{CONVEXBODY}, 
\code{POLYTOPE} or \code{ELLIPSOID}.}
}
\value{
Returns a vector.
}
\description{
Get a point on the boundary of a body.
}
