% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getCentre}
\alias{getCentre}
\alias{getCentre,ELLIPSOID-method}
\alias{getCentre,}
\alias{ELLIPSOID-method}
\title{getCentre}
\usage{
getCentre(object)

\S4method{getCentre}{ELLIPSOID}(object)
}
\arguments{
\item{object}{[ELLIPSOID] Any class that inherits from ELLIPSOID.}
}
\value{
Returns a matrix.
}
\description{
Get the centroid from an \code{ELLIPSOID} object.
}
