% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getEllipsoid}
\alias{getEllipsoid}
\alias{getEllipsoid,VBODY-method}
\alias{getEllipsoid,}
\alias{VBODY-method}
\title{getEllipsoid}
\usage{
getEllipsoid(object)

\S4method{getEllipsoid}{VBODY}(object)
}
\arguments{
\item{object}{[VBODY]}
}
\value{
Returns an object of class 'ELLIPSOID'.
}
\description{
Get ellipsoid object from a \code{VBODY} object.
}
