% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getHyperplane}
\alias{getHyperplane}
\alias{getHyperplane,POLYTOPE-method}
\alias{getHyperplane,}
\alias{POLYTOPEY-method}
\title{getHyperplane}
\usage{
getHyperplane(object)

\S4method{getHyperplane}{POLYTOPE}(object)
}
\arguments{
\item{object}{[POLYTOPE]}
}
\value{
Returns an object of class 'HYPERPLANE'.
}
\description{
Get hyperplane object from a \code{POLYTOPE} object.
}
