% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getInteriorPoint}
\alias{getInteriorPoint}
\alias{getInteriorPoint,ELLIPSOID-method}
\alias{getInteriorPoint,}
\alias{ELLIPSOID-method}
\alias{getInteriorPoint,POLYTOPE-method}
\alias{POLYTOPE-method}
\alias{getInteriorPoint,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{getInteriorPoint}
\usage{
getInteriorPoint(object)

\S4method{getInteriorPoint}{ELLIPSOID}(object)

\S4method{getInteriorPoint}{POLYTOPE}(object)

\S4method{getInteriorPoint}{CONVEXBODY}(object)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{CONVEXBODY}, 
\code{POLYTOPE} or \code{ELLIPSOID}.}
}
\value{
Returns a vector.
}
\description{
Get a point in the interior of a body.
}
