% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getNormal}
\alias{getNormal}
\alias{getNormal,HYPERPLANE-method}
\alias{getNormal,}
\alias{HYPERPLANE-method}
\title{getNormal}
\usage{
getNormal(object)

\S4method{getNormal}{HYPERPLANE}(object)
}
\arguments{
\item{object}{[HYPERPLANE] Any class that inherits from HYPERPLANE.}
}
\value{
Returns a matrix.
}
\description{
Get the normal vector from an \code{HYPERPLANE} object.
}
