% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getPolytope}
\alias{getPolytope}
\alias{getPolytope,VBODY-method}
\alias{getPolytope,}
\alias{VBODY-method}
\title{getPolytope}
\usage{
getPolytope(object)

\S4method{getPolytope}{VBODY}(object)
}
\arguments{
\item{object}{[VBODY]}
}
\value{
Returns an object of class 'VBODY'.
}
\description{
Get polytope object from a \code{VBODY} object.
}
