% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getRHS}
\alias{getRHS}
\alias{getRHS,POLYTOPE-method}
\alias{getRHS,}
\alias{POLYTOPE-method}
\alias{getRHS,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{getRHS,HYPERPLANE-method}
\alias{HYPERPLANE-method}
\title{getRHS}
\usage{
getRHS(object)

\S4method{getRHS}{POLYTOPE}(object)

\S4method{getRHS}{ELLIPSOID}(object)

\S4method{getRHS}{HYPERPLANE}(object)
}
\arguments{
\item{object}{[VBODY] Any class that inherits from VBODY.}
}
\value{
Returns a matrix.
}
\description{
Get the right-hand-side (RHS) variable from a \code{VBODY} object.
}
