% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getSamples}
\alias{getSamples}
\alias{getSamples,RP-method}
\alias{getSamples,}
\alias{RP-method}
\title{getSamples}
\usage{
getSamples(object, name = NULL)

\S4method{getSamples}{RP}(object, name = NULL)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{RP}.}

\item{name}{[character] Indicate which samples to return.}
}
\value{
Returns a matrix.
}
\description{
Get the samples from an object of class \code{RP}.
}
