% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getSense}
\alias{getSense}
\alias{getSense,VBODY-method}
\alias{getSense,}
\alias{VBODY-method}
\title{getSense}
\usage{
getSense(object)

\S4method{getSense}{VBODY}(object)
}
\arguments{
\item{object}{[VBODY] Any class that inherits from VBODY.}
}
\value{
Returns a matrix.
}
\description{
Get the right-hand-side (RHS) variable from a \code{VBODY} object.
}
