% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getShape}
\alias{getShape}
\alias{getShape,ELLIPSOID-method}
\alias{getShape,}
\alias{ELLIPSOID-method}
\title{getShape}
\usage{
getShape(object)

\S4method{getShape}{ELLIPSOID}(object)
}
\arguments{
\item{object}{[ELLIPSOID] Any class that inherits from ELLIPSOID.}
}
\value{
Returns a matrix.
}
\description{
Get the shape matrix from an \code{ELLIPSOID} object.
}
