% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_get.R
\docType{methods}
\name{getSpec}
\alias{getSpec}
\alias{getSpec,RP-method}
\alias{getSpec,}
\alias{RP-method}
\title{getSpec}
\usage{
getSpec(object)

\S4method{getSpec}{RP}(object)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{RP}.}
}
\value{
Returns a list
}
\description{
Get the spec slot from an object of class \code{RP}.
}
