% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{intersect}
\alias{intersect}
\alias{intersect,POLYTOPE,POLYTOPE-method}
\alias{intersect,}
\alias{POLYTOPE-method}
\title{intersect}
\usage{
intersect(x, y)

\S4method{intersect}{POLYTOPE,POLYTOPE}(x, y)
}
\arguments{
\item{x}{[S4 Class] Object that inherits from class \code{VBODY}.}

\item{y}{[S4 Class] Object that inherits from class \code{VBODY}.}
}
\value{
Returns an object containing the intersection of the 
        two input objects.
}
\description{
Creates an obejct
}
