% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{isEmpty}
\alias{isEmpty}
\alias{isEmpty,POLYTOPE-method}
\alias{isEmpty,}
\alias{POLYTOPE-method}
\alias{isEmpty,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\title{isEmpty}
\usage{
isEmpty(object)

\S4method{isEmpty}{POLYTOPE}(object)

\S4method{isEmpty}{ELLIPSOID}(object)
}
\arguments{
\item{object}{[POLYTOPE] Any class that inherits from POLYTOPE.}
}
\value{
Returns a matrix.
}
\description{
Checks if the input object is empty
}
