% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_miscellaneous.R
\name{lasserre}
\alias{lasserre}
\title{Lasserre's analytic solution}
\usage{
lasserre(a, th)
}
\description{
Implements Theorem 2.2 in Lasserre (2015)
}
\references{
J. B. Lasserre (2015). Volume of slices and section of the simplex in closed form
}
