% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Create a filename}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Numerical or integer indicates a year of a dataset}
}
\value{
Return a character string in a format "accident_<year>.csv.bz2" that
can be used as a filename
}
\description{
The function creates a filename for a .csv.bz2 based on the \code{year}
argument in a form "accident_<year>.csv.bz2".
It requires a numerical or integer input otherwise ends with an error.
}
\examples{
\dontrun{
make_filename(2014)
}
}
