% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R, R/rp_methods_set.R
\docType{class}
\name{POLYTOPE-class}
\alias{POLYTOPE-class}
\alias{addHyperplane<-,POLYTOPE-method}
\alias{replaceHyperplane<-,POLYTOPE-method}
\title{S4 class POLYTOPE}
\usage{
\S4method{addHyperplane}{POLYTOPE}(object) <- value

\S4method{replaceHyperplane}{POLYTOPE}(P, H, row_name = NULL) <- value
}
\description{
setClass call for object of class \code{POLYTOPE}.
}
