% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_set.R
\name{replaceHyperplane<-}
\alias{replaceHyperplane<-}
\title{Setter Function}
\usage{
replaceHyperplane(P, H, row_name) <- value
}
\arguments{
\item{object}{[POLYTOPE] object of class 'POLYTOPE'.}
}
\value{
Returns the modified input object.
}
\description{
Replacing an existing inequality.
}
