% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_classes.R, R/rp_methods_set.R
\docType{class}
\name{RP-class}
\alias{RP-class}
\alias{setSamples<-,RP-method}
\alias{setSpec<-,RP-method}
\alias{setPolytope<-,RP-method}
\alias{setEllipsoid<-,RP-method}
\alias{setConvexBody<-,RP-method}
\title{S4 class RP}
\usage{
\S4method{setSamples}{RP}(object) <- value

\S4method{setSpec}{RP}(object) <- value

\S4method{setPolytope}{RP}(object) <- value

\S4method{setEllipsoid}{RP}(object) <- value

\S4method{setConvexBody}{RP}(object) <- value
}
\description{
setClass call for object of class \code{RP}.
}
