% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{rp}
\alias{rp}
\alias{rp,SIMPLEX-method}
\alias{rp,}
\alias{SIMPLEX-method}
\alias{rp,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{rp,POLYTOPE-method}
\alias{POLYTOPE-method}
\alias{rp,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{rp}
\usage{
rp(object, spec)

\S4method{rp}{SIMPLEX}(object, spec)

\S4method{rp}{ELLIPSOID}(object, spec)

\S4method{rp}{POLYTOPE}(object, spec)

\S4method{rp}{CONVEXBODY}(object, spec)
}
\arguments{
\item{object}{[CONVEXBODY] Any class that inherits from CONVEXBODY.}

\item{spec}{[list]  A list of specifications passed to lower level functions.}
}
\value{
Returns a matrix.
}
\description{
Random sampling from object
}
