% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods_set.R
\name{setEllipsoid<-}
\alias{setEllipsoid<-}
\title{Setter Function}
\usage{
setEllipsoid(object) <- value
}
\arguments{
\item{object}{[RP] object of class 'RP'.}
}
\value{
Returns the modified input object.
}
\description{
Set ellipsoid to object.
}
