% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{thin}
\alias{thin}
\alias{thin,RP-method}
\alias{thin,}
\alias{RP-method}
\title{thin}
\usage{
thin(object)

\S4method{thin}{RP}(object)
}
\arguments{
\item{object}{[S4] Any class that inherits from class \code{RP}.}
}
\value{
Returns the thinned input object.
}
\description{
Thins the samples of an RP object.
}
