% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rp_methods.R
\docType{methods}
\name{transform}
\alias{transform}
\alias{transform,RPP-method}
\alias{transform,}
\alias{RPP-method}
\alias{transform,RPE-method}
\alias{RPE-method}
\alias{transform,RPB-method}
\alias{transform,POLYTOPE-method}
\alias{POLYTOPE-method}
\alias{transform,ELLIPSOID-method}
\alias{ELLIPSOID-method}
\alias{transform,CONVEXBODY-method}
\alias{CONVEXBODY-method}
\title{transform}
\usage{
transform(object)

\S4method{transform}{RPP}(object)

\S4method{transform}{RPE}(object)

\S4method{transform}{RPB}(object)

\S4method{transform}{POLYTOPE}(object)

\S4method{transform}{ELLIPSOID}(object)

\S4method{transform}{CONVEXBODY}(object)
}
\arguments{
\item{object}{[S4 Class] Object of class \code{VBODY}, \code{RP}.}
}
\value{
Returns an object of the same class as the imput 
        but transformed to (d-1)-dimensional space.
}
\description{
Perrforms a (isometric, i.e. distance preserving) transformation
             of body from d-demensional space to (d-1)-demensional space.
             Note that the body needs to be a simplex or a the intersection of a 
             body with a simplex (i.e. a constrained simplex).
}
